Add transactions with splits and demonstrate bug with stock purchases
In the "Buy Stock" transaction that was added in this commit, notice that one of the splits is
marked as `"JPY"  0`. On the GUI, I can see that this is `"INR"  10`. However, this does not get
translated correctly because the "Edit Exchange Rate" dialog box is not shown by GnuCash within the
a stock purchase register!